feat: åpne for å vise PortalImage i en Grid
affects: @fremtind/portal
