+++ portal/src/components/portal-image/portal-image.scss
